feat: introduces a "server" package and moves UmbAppContext logic to UmbServerContext
