+++ src/Umbraco.Web.UI.Client/src/packages/core/server/network-connection-status.manager.ts
