+++ src/Umbraco.Web.UI.Client/src/packages/core/server/server.context-token.ts
@@ -1,0 +1,4 @@
+import type { UmbServerContext } from './server.context.js';
+import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+
+export const UMB_SERVER_CONTEXT = new UmbContextToken<UmbServerContext>('UmbServerContext');
