+++ src/Umbraco.Web.UI.Client/tsconfig.json
@@ -117,6 +117,7 @@ DON'T EDIT THIS FILE DIRECTLY. It is generated by /devops/tsconfig/index.js
 			"@umbraco-cms/backoffice/script": ["./src/packages/templating/scripts/index.ts"],
 			"@umbraco-cms/backoffice/search": ["./src/packages/search/index.ts"],
 			"@umbraco-cms/backoffice/section": ["./src/packages/core/section/index.ts"],
+			"@umbraco-cms/backoffice/server": ["./src/packages/core/server/index.ts"],
 			"@umbraco-cms/backoffice/server-file-system": ["./src/packages/core/server-file-system/index.ts"],
 			"@umbraco-cms/backoffice/settings": ["./src/packages/settings/index.ts"],
 			"@umbraco-cms/backoffice/sorter": ["./src/packages/core/sorter/index.ts"],
